포인트 조회api

+++ src/main/java/com/cdy/ecommerce/ecommerce/api/admin/product/ProductAdminDTO.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 
 public class ProductAdminDTO {
 
@@ -16,7 +18,7 @@ public class ProductAdminDTO {
     public static class Request {
         private String userId; // 상품을 등록하는 사용자의 ID
         private String name; // 상품 이름
-        private int price; // 상품 가격
+        private BigDecimal price; // 상품 가격
         private int initialStock; // 초기 재고 수량
     }
 
@@ -27,7 +29,7 @@ public class ProductAdminDTO {
     public static class Response {
         private Long id; // 상품 ID
         private String name; // 상품 이름
-        private int price; // 상품 가격
+        private BigDecimal price; // 상품 가격
         private int stock; // 현재 재고 수량
     }
 }
