+++ src/main/java/com/cdy/ecommerce/ecommerce/api/v1/order/controller/ProductOrderDTO.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class ProductOrderDTO {
@@ -28,7 +29,7 @@ public class ProductOrderDTO {
         private Long orderId;
         private String status;
         private List<OrderItem> items;
-        private int totalAmount;
+        private BigDecimal totalAmount;
     }
 
     @Getter
@@ -36,7 +37,7 @@ public class ProductOrderDTO {
     public static class OrderItem {
         private Long productId;
         private int quantity;
-        private int price;
+        private BigDecimal price;
     }
 
 }
