+++ src/main/java/com/cdy/ecommerce/ecommerce/api/v1/product/controller/ProductDTO.java
@@ -2,6 +2,8 @@ package com.cdy.ecommerce.ecommerce.api.v1.product.controller;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductDTO {
 
@@ -13,7 +15,7 @@ public class ProductDTO {
 
     private String userId;
     private String name;
-    private int price;
+    private BigDecimal price;
     private int stock;
   }
 
@@ -24,7 +26,7 @@ public class ProductDTO {
   public static class Response {
     private Long id;
     private String name;
-    private int price;
+    private BigDecimal price;
     private long stock;
   }
 }
