+++ src/test/java/com/cdy/ecommerce/usecase/stub/ProductSteps.java
@@ -2,12 +2,14 @@ package com.cdy.ecommerce.usecase.stub;
 
 import com.cdy.ecommerce.ecommerce.api.v1.product.controller.ProductDTO;
 
+import java.math.BigDecimal;
+
 
 public class ProductSteps {
     public static ProductDTO.Request addProductRequest() {
         final String userId = "user0";
         final String name = "상품명";
-        final int price = 1000;
+        final BigDecimal price = BigDecimal.valueOf(1000);
         final int stock = 100;
         return new ProductDTO.Request(userId, name, price, stock);
     }
